requirements.txt
Pyrogram
TgCrypto
aiohttp
hurry.filesize